delete duplicate scripts
