+++ packages/splits/package.json
@@ -33,15 +33,12 @@
     "@cobuild/utility-contracts": "workspace:*",
     "@openzeppelin/contracts": "5.0.0",
     "@openzeppelin/contracts-upgradeable": "5.0.0",
-    "@wagmi/cli": "^1.5.2",
     "ds-test": "https://github.com/dapphub/ds-test#cd98eff28324bfac652e63a239a60632a761790b",
     "forge-std": "https://github.com/foundry-rs/forge-std#87a2a0afc5fafd6297538a45a52ac19e71a84562",
     "solmate": "^6.1.0"
   },
   "devDependencies": {
     "@types/node": "^20.10.4",
-    "glob": "^10.3.10",
-    "pathe": "^1.1.1",
     "prettier": "2.8.8",
     "prettier-plugin-solidity": "1.2.0",
     "solhint": "3.4.1",
